requirements.txt
streamlit
pandas
requests
matplotlib
scikit-learn
python-dotenv
streamlit-autorefresh
plotly
smtplib